pyproject.toml
[project]
name = "awspythonargparse"
version = "0.1.0"
description = ""
authors = [
    {name = "Zuka Esakia",email = "A.Esakia@smartsoft.ge"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3 (>=1.38.18,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "awspythonargparse", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
